fixes build scripts for mac

+++ stage3/04-install-sw/01-run.sh
@@ -2,7 +2,7 @@
 
 # Set up .pdsettings link for root user, so the same settings are used for normal and root user.
 on_chroot << EOF
-	ln -s /home/${FIRST_USER_NAME}/.pdsettings /root/.pdsettings
+	ln -sfn /home/${FIRST_USER_NAME}/.pdsettings /root/.pdsettings
 EOF
 
 # Copy preconfigured PD settings.
@@ -17,7 +17,9 @@ mkdir -p "${ROOTFS_DIR}/home/${FIRST_USER_NAME}/Desktop"
 cp files/Desktop/*.desktop "${ROOTFS_DIR}/home/${FIRST_USER_NAME}/Desktop/"
 
 # For unknown reason SuperColliderIDE.desktop does not show up in system menu, unless it's renamed...
-mv "${ROOTFS_DIR}/usr/share/applications/SuperColliderIDE.desktop" "${ROOTFS_DIR}/usr/share/applications/SuperCollider.desktop"
+if [ -f "${ROOTFS_DIR}/usr/share/applications/SuperColliderIDE.desktop" ]; then
+	mv "${ROOTFS_DIR}/usr/share/applications/SuperColliderIDE.desktop" "${ROOTFS_DIR}/usr/share/applications/SuperCollider.desktop"
+fi
 
 # Copy preconfigured Pianoteq settings.
 mkdir -p "${ROOTFS_DIR}/home/${FIRST_USER_NAME}/.config/Modartt"
